修复周期记账、数据库升级

+++ SuiShouJi/Components/Business/DataSync/SSJDataSynchronizeTask.m
@@ -110,7 +110,7 @@ static NSString *const kSyncZipFileName = @"sync_data.zip";
                 NSString *desc = responseInfo[@"desc"];
                 tError = [NSError errorWithDomain:SSJErrorDomain code:code userInfo:@{NSLocalizedDescriptionKey:desc}];
                 
-                SSJPRINT(@">>> SSJ Wanings:同步失败-----code:%d desc:%@", code, desc);
+                SSJPRINT(@">>> SSJ Wanings:同步失败-----code:%ld desc:%@", (long)code, desc);
                 //                SSJPRINT(@">>> SSJ sync response data:%@", responseObject);
                 if (failure) {
                     failure(tError);
